+++ src/output/characterManager/CharacterManager.cpp
@@ -1,0 +1,63 @@
+#include "CharacterManager.h"
+
+CharacterManager* characterManagerPtr;
+
+
+CharacterManager::CharacterManager()
+{
+    characterManagerPtr = this;
+}
+
+void CharacterManager::GiveUpgrade(const Upgrades upgrade)
+{
+    if (Current_CharObj2 != nullptr)
+        Current_CharObj2->Upgrades |= upgrade;
+    if (CharObj2Ptrs != nullptr)
+    {
+        for (const auto& character : CharObj2Ptrs)
+        {
+            if (character != nullptr)
+                character->Upgrades |= upgrade;
+        }
+    }
+}
+
+
+void CharacterManager::RemoveUpgrade(const Upgrades upgrade)
+{
+    if (Current_CharObj2 != nullptr)
+        Current_CharObj2->Upgrades &= ~upgrade;
+
+    for (const auto& character : CharObj2Ptrs)
+    {
+        if (character != nullptr)
+            character->Upgrades &= ~upgrade;
+    }
+}
+
+static void __cdecl HandleRingLoss()
+{
+    switch (characterManagerPtr->options.ringLoss)
+    {
+    case Classic:
+        Set0Rings();
+        break;
+    case Modern:
+        if (Rings > 20)
+            Rings -= 20;
+        else
+            Set0Rings();
+        break;
+    case OneHitKnockOut:
+        Set0Rings();
+        KillHimP(0);
+        break;
+    }
+}
+
+void CharacterManager::UpdateOptions(const Options newOptions)
+{
+    this->options = newOptions;
+    //We override the Set0Rings call inside the HurtPlayer function;
+    WriteCall(reinterpret_cast<void*>(0x45072D), &HandleRingLoss);
+}